requirements.txt
Django==3.2
django-cleanup==5.2.0
djangorestframework==3.12.4
mysqlclient==2.0.3
Pillow==8.2.0
graphene-django==2.15.0
python-decouple==3.4
gunicorn==20.1.0